worker-python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
numpy>=1.26.0
sentence-transformers>=3.0.0
